Swig wrappers for OpenSceneGraph.
This is the new trunk of osgswig.

 http://osgswig.googlecode.com/svn/trunk
 or for members: https://osgswig.googlecode.com/svn/trunk

it is a merge of the original osgswig branch and
the branch used for the VRmeer Project (python and perl).

The merge contains new cmake files, that build the
following modules:

osg
osgDB
osgFX
osgGA
osgUtil
osgViewer

There is one (new) simple test example:
examples/python/simple/simpleosg

The default cmakesetting make these modules for python.

It is most recently tested with
swig 1.3.33
OpenSceneGraph 2.3.4

The merge also still contain all the files and documentation from the original osgswig.
revision 95 of:
 http://osgswig.googlecode.com/svn/
 or https://osgswig.googlecode.com/svn/ (for members)
contains the original osgswig.